# speed up bam.Read