This was an experiment to see if it would improve bam parsing speed.
It replaces the smaller reads with a single, large read and then just works off the []byte.

With Omit(0) on an example bam, I get 15.5 seconds with this change and 18.3 on master.
With 3 cpus, I get 7.9 secs here and 11.3 master.
For Omit > 0, the difference, as expected is smaller, but still present.

The `TestConceptualBAM` test is failing. I'm not sure how to address that.