@kortschak 2016-10-23
I can't accept this until the conceptual bam test passes. That test behaviour that is necessary for BAM indexing.