@kortschak 2016-10-23
I don't think this is the right way to do it.

What you are doing is intended to do a buffered read of the record. This work all ends up in the `Read` method, but it is not necessary for this to work. All you need to do is add a buffer to `errReader` that you can preload with `block_size` bytes from the `bgzf.Reader` after you have read the first four bytes. This will leave the body of `Read` essentially unchanged (the nature of errors will be difference - unfortunately losing the context information content of the errors).

There are also other potential target for optimisation: `binReader` should may be improved by not having methods on the pointer and there may be a benefit from using a `sync.Pool` to keep the readers (thus saving the buffer allocations).