@brentp 2016-10-23
I did like this approach because it is transparent and even removes lots of code. Would you consider it if the tests pass?

Otherwise, you mean

```
type errReader struct {
    r   *bgzf.Reader
    n   int
    err error
    buf [bgzf.MaxBlockSize]
}
```
?
and to Read from that and return it to a sync.Pool?