@kortschak 2016-10-23
The change moves code from a relatively regular and simple section into the spec-dependent section. This makes `Read` more difficult to vet without significantly improving the readability overall.

Yes, essentially like that, though the buffer does not need to be BGZF block size large. Probably a buffer of 4096 would be fine - a `bytes.Buffer` would work except that it reads ahead and that would break indexing. So essentially you need to mimic a `bytes.Buffer` with a read-ahead constraint.