@brentp 2016-10-24
I gave this a try quickly here: https://github.com/brentp/hts/commit/b46df5abaecff103ff59c5b953d2cbab682e27e8
using a limited reader, a [1024]byte buffer in errReader, and a sync.Pool to reuse. 

On a bam where no blocks exceed 256. I get:

```
limited-reader
real	0m9.610s
user	0m25.300s
sys	0m0.340s

master
real	0m11.319s
user	0m27.040s
sys	0m0.344s

one-read
real	0m7.849s
user	0m24.388s
sys	0m0.392s
```

Also, the limited-reader branch still fails the ConceptualBam test.