@kortschak 2016-10-24
That looks like it's heading in the right direction, though pools should be shared across all values in the package, not just be a single reader (there should be a helper for get and put to make sure the state is correct - see how it's done in gonum/matrix/mat64 for the types there).

There is unnecessary copying in the code as it stands (this is because of what I have asked you to do), but this can go away later if the errReader is made to be more specialised and be able to return binary decodings directly. I suspect that after that we will see similar performance as with one-read.

However, the indexing behaviour needs to be fixed.