@brentp 2016-11-01
do you have any insight on why the TestConceptualBam is failing or how to fix? I see that on master, it never reaches the final block, instead, it sends an EOF *with* the penultimate block.

In master, it reads data and fills the sam record for the block, but then checks r.err right after parseAux() [here](https://github.com/biogo/hts/blob/50da7d4131a3b5c9d063932461cab4d1fafb20b0/bam/reader.go#L181). And it returns nil, r.err there, even though there is (what appears to be) valid data in the sam.Record.


IIUC, the reason it gets an EOF there is because `len(auxTags) == 0` and it tries to read 0 bytes. Which causes it to skip to the next bgzf block. but that still discards a valid record. In short, I think the test is flawed.