@kortschak 2016-11-01
I think you are right.

I'll send a PR to fix that.

Did you notice that your PR panics on 1.3 and 1.4? I'm happy to drop support for those Go versions, but do you know why?