@brentp 2016-11-02
hmm, I d/l'ed 1.4.3 to try it out. If I replace:
```
c.raw[r.off:r.plus(4)]
```
with 
```
v := c.raw[r.off:r.off + 4]
r.off += 4
```
thenblob/master/doc/go1.5.html it works in 1.4. so must be an order of operations change. I can change the read* methods to be:
```
s, e := r.off, r.plus(n)
c.raw[s:e]
```
what do you think?