@kortschak 2016-11-02
TBH I'd like the plus API to go away. I'd prefer if you had a something that returned the bytes and updated the position it would be nicer. Then there is no need to worry about this.