@brentp 2016-11-02
something like this: 
```
func (c *cursor) get(n int) []byte {
    s, e := c.off, c.off + n
    c.off = e
    return c.raw[s:e]
 }
```