@kortschak 2016-11-02
Sure, though s/get/bytes/.