@brentp 2016-11-02
Done. Any other concerns?

now the tests are failing with `gofmt` not found.