@kortschak 2016-11-02
That's not gofmt not found - that warning always happens. However, your code is not gofmt'd.

I'd like to get #39 merged first and then rebase this on top of that as a single commit if that's OK. Then I'll have a more careful look here.