@brentp 2016-11-03
rebased and ready for review.