@brentp 2016-11-03
I pushed a change that makes all other changes except for using bgzf.Tx which I'm not familiar with.
I will squash after you take a look.