@brentp 2016-11-03
current state is right around 8 seconds for the same test I posted before with master at 11.3