@kortschak 2016-11-03
Yes, this looks OK.

For the Tx stuff, you end up with a newBuffer that looks like this (untested):

```
func newBuffer(r *bgzf.Reader) (*buffer, bgzf.Chunk, error) {
	var buf [4]byte
	b := &buffer{data: buf[:4]}
	size := int(b.readInt32())
	tx := r.Begin()
	b.off, b.data = 0, make([]byte, size)
	n, err := r.Read(b.data)
	chunk := tx.End()
	if err != nil {
		return nil, chunk, err
	}
	if n != size {
		return nil, chunk, errors.New("bam: truncated record")
	}
	return b, chunk, nil
}
```

It may be better to have the [4]byte as part of the buffer, but we can try that later.