@brentp 2016-11-03
I have this:
```
  func newBuffer(br *Reader, buf [4]byte) (*buffer, error) {
  
      b := &buffer{data: buf[:4]}
      size := int(b.readInt32())
      tx := br.r.Begin()
      b.off, b.data = 0, make([]byte, size)
      n, err := br.r.Read(b.data)
      br.lastChunk = tx.End()
      if err != nil {
          return nil, err
      }
      if n != size {
          return nil, errors.New("bam: truncated record")
      }
      return b, nil
  }
```