@brentp 2016-11-04
In actual tests, I see 30-35% improvement from this. [A benchmark](https://github.com/brentp/biogo-bench) shows even better improvement:

```
brentp@x1:~/go/src/github.com/brentp/biogo-bench$ benchcmp before.txt after.txt 
benchmark           old ns/op     new ns/op     delta
BenchmarkRead-4     346           135           -60.98%

benchmark           old allocs     new allocs     delta
BenchmarkRead-4     3              1              -66.67%

benchmark           old bytes     new bytes     delta
BenchmarkRead-4     224           4             -98.21%
```

where before is d3530d4