@brentp 2016-11-04
[here's](https://github.com/brentp/biogo-bench) a comparison to d3530d4

```
benchmark           old ns/op     new ns/op     delta
BenchmarkRead-4     8711209       6306822       -27.60%

benchmark           old allocs     new allocs     delta
BenchmarkRead-4     40688          34018          -16.39%

benchmark           old bytes     new bytes     delta
BenchmarkRead-4     4076490       4388706       +7.66%
```